requirements.txt
Flask
Flask-RESTful
Flask-Cors
mistune
PyYAML
